extracts encryption methods into a module and refactors valid? method
