+++ p_sets/2/caesar_test.rb
@@ -1,8 +1,7 @@
 require 'minitest/autorun'
 require 'minitest/reporters'
-reporter_options = { color: true }
-Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new(reporter_options)]
 require_relative 'caesar'
+Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new({color: true})]
 
 class CaesarProgramTest < Minitest::Test
   def setup
