+++ p_sets/2/vigenere_test.rb
@@ -1,8 +1,8 @@
 require 'minitest/autorun'
 require 'minitest/reporters'
-reporter_options = { color: true }
-Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new(reporter_options)]
 require_relative 'vigenere'
+
+Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new({color: true})]
 
 class VigenereTestProgram < Minitest::Test
   def setup
